Make assertInvalid* delegates to functionAssertions
